+++ app/test/library_controller_test.dart
@@ -1387,6 +1387,7 @@ class FakeFilePickerPlatform extends FilePickerPlatform {
     Function(FilePickerStatus)? onFileLoading,
     int compressionQuality = 0,
     AndroidOptions androidOptions = const AndroidOptions(),
+    DarwinOptions darwinOptions = const DarwinOptions(),
     WindowsOptions windowsOptions = const WindowsOptions(),
     LinuxOptions linuxOptions = const LinuxOptions(),
     WebOptions webOptions = const WebOptions(),
@@ -1433,6 +1434,9 @@ base class _FakePlatformFile extends PlatformFile {
 
   @override
   Future<int> length() async => size;
+
+  @override
+  int? lengthSync() => size;
 
   @override
   Future<Uint8List> readAsBytes() async => bytes;